+++ lib/turbo_firestore_api.dart
@@ -79,6 +79,8 @@ export 'mixins/turbo_exception_handler.dart';
 /// Type definitions
 export 'typedefs/collection_reference_def.dart';
 export 'typedefs/create_doc_def.dart';
+export 'typedefs/update_doc_def.dart';
+export 'typedefs/upsert_doc_def.dart';
 export 'typedefs/turbo_locator_def.dart';
 
 /// Constants