+++ lib/typedefs/update_doc_def.dart
@@ -1,4 +1,3 @@
 import 'package:turbo_firestore_api/models/turbo_auth_vars.dart';
 
 typedef UpdateDocDef<T> = T Function(T current, TurboAuthVars vars);
-typedef NullableUpdateDocDef<T> = T? Function(T? current, TurboAuthVars vars);