+++ lib/typedefs/upsert_doc_def.dart
@@ -1,0 +1,3 @@
+import 'package:turbo_firestore_api/models/turbo_auth_vars.dart';
+
+typedef UpsertDocDef<T> = T Function(T? current, TurboAuthVars vars);